+++ frontend/src/components/IndividualThearapy.js
@@ -1,6 +1,6 @@
 import React, {useState} from "react";
 import { Link } from "react-router-dom";
-import { checkToken } from "../utils/checkToken";
+import { checkToken } from "../utils/Hooks/checkToken";
 
 const IndividualThearapy = () => {
   return (
