--- frontend/src/components/TherapistHome.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function TherapistHome() {
-  return (
-    <div>
-      comming soon TherapistHome
-    </div>
-  )
-}
-
-export default TherapistHome
