+++ frontend/src/components/User/BookTherapistPage.js
@@ -1,9 +1,10 @@
 import React, { useEffect, useState } from "react";
-import Side_Navbar from "./Side_Navbar";
+import Side_Navbar from "../Side_Navbar";
 import { useDispatch, useSelector } from "react-redux";
 import { Link, useParams } from 'react-router-dom';
-import VerifyClient from "../utils/verifyClient";
-import { addUser } from "../utils/userSlice";
+
+import { addUser } from "../../utils/Slices/userSlice";
+import VerifyClient from "../../utils/Hooks/verifyClient";
 
 function BookTherapistPage() {
   const [IsAuthorized, setIsAuthorized] = useState(true);
