+++ frontend/src/utils/Hooks/checkToken.js
@@ -1,3 +1,3 @@
 export const checkToken = (tokenName) => {
-  return !!localStorage.get(tokenName);
+  return !!localStorage.getItem(tokenName);
 };
